Pointing to a before commit status
