+++ crates/cairo-lint-core/src/lints/mod.rs
@@ -2,6 +2,6 @@ pub mod bool_comparison;
 pub mod breaks;
 pub mod double_comparison;
 pub mod double_parens;
-pub mod equal_operand;
+pub mod ifs;
 pub mod loops;
 pub mod single_match;
