+++ crates/cairo-lint-core/src/plugin.rs
@@ -3,11 +3,11 @@ use cairo_lang_defs::plugin::PluginDiagnostic;
 use cairo_lang_semantic::db::SemanticGroup;
 use cairo_lang_semantic::plugin::{AnalyzerPlugin, PluginSuite};
 use cairo_lang_semantic::Expr;
-use cairo_lang_syntax::node::ast::Expr as AstExpr;
+use cairo_lang_syntax::node::ast::{Expr as AstExpr, ExprBinary};
 use cairo_lang_syntax::node::kind::SyntaxKind;
 use cairo_lang_syntax::node::{TypedStablePtr, TypedSyntaxNode};
 
-use crate::lints::{bool_comparison, breaks, double_comparison, double_parens, loops, single_match};
+use crate::lints::{bool_comparison, breaks, double_comparison, double_parens, ifs, loops, single_match};
 
 pub fn cairo_lint_plugin_suite() -> PluginSuite {
     let mut suite = PluginSuite::default();
@@ -39,7 +39,7 @@ pub fn diagnostic_kind_from_message(message: &str) -> CairoLintKind {
         double_comparison::CONTRADICTORY_COMPARISON => CairoLintKind::DoubleComparison,
         breaks::BREAK_UNIT => CairoLintKind::BreakUnit,
         bool_comparison::BOOL_COMPARISON => CairoLintKind::BoolComparison,
-        equal_operand::REDUNDANT_COMPARISON => CairoLintKind::RedundantComparison,
+        ifs::REDUNDANT_COMPARISON => CairoLintKind::RedundantComparison,
         _ => CairoLintKind::Unknown,
     }
 }
@@ -63,7 +63,8 @@ impl AnalyzerPlugin for CairoLint {
                         loops::check_loop_match_pop_front(db, expr_loop, &mut diags, &function_body.arenas)
                     }
                     Expr::If(expr_if) => {
-                        equal_operand::check_redundant_comparison(db, expr_if, &mut diags, &function_body.arenas)
+                        print!("EXPRESSION: {:?}", expr_if);
+                        ifs::check_redundant_comparison(db, expr_if, &mut diags, &function_body.arenas)
                     }
                     _ => (),
                 };
